\docType{package}
\name{antiProfiles-package}
\alias{antiProfiles-package}
\title{Gene Expression Anti-profiles}
\description{
  This package implements the anti-profile method of
  Corrada Bravo et al., BMC Bioinformatics 2012, 13:272
  doi:10.1186/1471-2105-13-272.
}
\author{
  Hector Corrada Bravo \email{hcorrada@gmail.com}
}

